migrate save related reducers from desktop to core
